zephyr/motionsense: ensure default_range is signed
Not casting the value of default_range from the device tree to a signed
value causes clang to raise a warning (promoted to an error) regarding
narrowing conversions if using ROUND_UP_FLAG. ROUND_UP_FLAG is defined
to be an unsigned value, so ROUND_UP_FLAG | <literal integer> is
interpreted as unsigned which then causes a warning because
the resulting value is converted from a large positive number to a
large negative one.

It would generally be more correct to make the range unsigned, but it's
difficult to ensure that correct behavior is retained because some code
paths do arithmetic on signed sensor readings and changing the range
fields of struct motion_sensor_t can have surprising effects on those
operations. Multiplying or dividing a signed number (like a sensor
reading) by an unsigned one (the range) causes the signed value to be
promoted to an unsigned one first which will lead to anomalously large
computed values anytime the input value is negative.

BUG=None
TEST=twister

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4583469
Tested-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>

+++ zephyr/shim/src/motionsense_sensors.c
@@ -237,7 +237,7 @@ DT_FOREACH_STATUS_OKAY(cros_ec_motionsense_bodydetect,
 	.name = DT_NODE_FULL_NAME(id),                                       \
 	.active_mask = DT_STRING_TOKEN(id, active_mask),                     \
 	.location = DT_STRING_TOKEN(id, location),                           \
-	.default_range = DT_PROP(id, default_range),                         \
+	.default_range = (int)(DT_PROP(id, default_range)),                  \
 	SENSOR_I2C_SPI_ADDR_FLAGS(id) SENSOR_MUTEX(id) SENSOR_I2C_PORT(id)   \
 		SENSOR_ROT_STD_REF(id) SENSOR_DRV_DATA(id) SENSOR_CONFIG(id) \
 			SENSOR_INT_SIGNAL(id) SENSOR_FLAGS(id)               \
